Splited ShareInvoker class file Usage new ShareInvoker(Activity.this, mMap).shareMapSnapshot()
