+++ app/src/main/java/com/canvearth/canvearth/MapsActivity.java
@@ -1,16 +1,12 @@
 package com.canvearth.canvearth;
 
 import android.Manifest;
-import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.location.Location;
-import android.net.Uri;
 import android.os.AsyncTask;
-import android.provider.MediaStore;
 import android.support.annotation.NonNull;
 import android.support.v4.app.ActivityCompat;
 import android.os.Bundle;
@@ -29,7 +25,7 @@ import com.canvearth.canvearth.pixel.Pixel;
 import com.canvearth.canvearth.utils.Constants;
 import com.canvearth.canvearth.utils.PermissionUtils;
 import com.canvearth.canvearth.utils.PixelUtils;
-import com.facebook.login.LoginManager;
+import com.canvearth.canvearth.utils.ShareInvoker;
 import com.github.pengrad.mapscaleview.MapScaleView;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.OnMapReadyCallback;
@@ -37,7 +33,6 @@ import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.CameraPosition;
 
 import java.io.InputStream;
-import java.util.Arrays;
 
 public class MapsActivity extends AppCompatActivity implements
         OnMapReadyCallback,
@@ -79,40 +74,6 @@ public class MapsActivity extends AppCompatActivity implements
         }
     }
 
-    protected class ShareInvoker extends AsyncTask<Void, Void, Void>{
-        protected Activity mParentActivity;
-        protected GoogleMap mMap;
-
-        public ShareInvoker(Activity parentActivity, GoogleMap map) {
-            this.mParentActivity = parentActivity;
-            this.mMap = map;
-        }
-
-        @Override
-        protected Void doInBackground(Void... voids) {
-            if (mParentActivity.checkSelfPermission(Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
-                mMap.snapshot(new GoogleMap.SnapshotReadyCallback() {
-                    @Override
-                    public void onSnapshotReady(Bitmap bitmap) {
-                        String bitmapPath = MediaStore.Images.Media.insertImage(getContentResolver(), bitmap, "canvearthMapSnapshot", null);
-                        Uri bitmapUri = Uri.parse(bitmapPath);
-
-                        Log.d(TAG, "snapshot Ready");
-                        Intent sendIntent = new Intent();
-                        sendIntent.setAction(Intent.ACTION_SEND)
-                                .setType("image/*")
-                                .putExtra(Intent.EXTRA_STREAM, bitmapUri);
-                        startActivity(sendIntent);
-                    }
-                });
-            }
-            else {
-                ActivityCompat.requestPermissions(mParentActivity, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, 1);
-            }
-            return null;
-        }
-    }
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -132,7 +93,7 @@ public class MapsActivity extends AppCompatActivity implements
         shareButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                new ShareInvoker(MapsActivity.this, mMap).execute();
+                new ShareInvoker(MapsActivity.this, mMap).shareMapSnapshot();
             }
         });
